extracted refinements
