+++ src/Language/Fixpoint/Types.hs
@@ -26,6 +26,7 @@ module Language.Fixpoint.Types (
   , module Language.Fixpoint.Types.Names
   , module Language.Fixpoint.Types.Sorts
   , module Language.Fixpoint.Types.Refinements
+  , module Language.Fixpoint.Types.Substitutions
 
   , toFixpoint
   , writeFInfo
@@ -69,30 +70,6 @@ module Language.Fixpoint.Types (
   -- * Qualifiers
   , Qualifier (..)
 
-  -- * Refinements
-  , SortedReft (..), Reft(..), Reftable(..)
-
-  -- * Constructing Refinements
-  , reft                    -- "smart
-  , trueSortedReft          -- trivial reft
-  , trueReft                -- trivial reft
-  , exprReft                -- singleton: v == e
-  , notExprReft             -- singleton: v /= e
-  , uexprReft               -- singleton: v ~~ e
-  , symbolReft              -- singleton: v == x
-  , usymbolReft             -- singleton: v ~~ x
-  , propReft                -- singleton: Prop(v) <=> p
-  , predReft                -- any pred : p
-  , reftPred, reftBind
-  , isFunctionSortedReft, functionSort
-  , isNonTrivial
-  , isSingletonReft
-  , isEVar
-  , isFalse
-  , flattenRefas, conjuncts
-  , shiftVV
-  , mapPredReft
-
   -- * Cut KVars
   , Kuts (..)
   , ksEmpty
@@ -103,64 +80,37 @@ module Language.Fixpoint.Types (
   , convertFormat
   ) where
 
-import           Debug.Trace               (trace)
+-- import           Debug.Trace               (trace)
 
 import qualified Data.Binary as B
 import           Data.Generics             (Data)
 import           Data.Typeable             (Typeable)
 import           GHC.Generics              (Generic)
-
 import           Data.Hashable
-import           Data.List                 (partition, foldl', sort, sortBy)
-import           Data.String
+import           Data.List                 (partition, sort)
+-- import           Data.String
 import           Data.Text                 (Text)
 import qualified Data.Text                 as T
 import           Control.DeepSeq
 import           Data.Maybe                (isJust, mapMaybe, listToMaybe, fromMaybe)
-import           Text.Printf               (printf)
 import           Language.Fixpoint.Types.Config
 import           Language.Fixpoint.Types.Names
 import           Language.Fixpoint.Types.PrettyPrint
 import           Language.Fixpoint.Types.Errors
 import           Language.Fixpoint.Types.Spans
 import           Language.Fixpoint.Types.Sorts
+import           Language.Fixpoint.Types.Refinements
+import           Language.Fixpoint.Types.Substitutions
 import           Language.Fixpoint.Utils.Misc
-import           Text.Parsec.Pos
+-- import           Text.Parsec.Pos
 import           Text.PrettyPrint.HughesPJ
 
-import           Data.Array                hiding (indices)
+-- import           Data.Array                hiding (indices)
 import qualified Data.HashMap.Strict       as M
 import qualified Data.HashSet              as S
 
-exprSymbols :: Expr -> [Symbol]
-exprSymbols = go
-  where
-    go (EVar x)        = [x]
-    go (EApp f es)     = val f : concatMap go es
-    go (ENeg e)        = go e
-    go (EBin _ e1 e2)  = go e1 ++ go e2
-    go (EIte p e1 e2)  = predSymbols p ++ go e1 ++ go e2
-    go (ECst e _)      = go e
-    go _               = []
-
-predSymbols :: Pred -> [Symbol]
-predSymbols = go
-  where
-    go (PAnd ps)          = concatMap go ps
-    go (POr ps)           = concatMap go ps
-    go (PNot p)           = go p
-    go (PIff p1 p2)       = go p1 ++ go p2
-    go (PImp p1 p2)       = go p1 ++ go p2
-    go (PBexp e)          = exprSymbols e
-    go (PAtom _ e1 e2)    = exprSymbols e1 ++ exprSymbols e2
-    go (PKVar _ (Su su))  = {- CUTSOLVER k : -} syms (M.keys su) ++ syms (M.elems su)
-    go (PAll xts p)       = (fst <$> xts) ++ go p
-    go _                  = []
-
 newtype Kuts = KS { ksVars :: S.HashSet KVar }
                deriving (Eq, Show, Generic)
-
-
 
 instance Fixpoint Kuts where
   toFix (KS s) = vcat $ ((text "cut " <>) . toFix) <$> S.toList s
@@ -178,266 +128,10 @@ ksMember k (KS s) = S.member k s
 -- | Type Constructors ---------------------------------------------------------
 --------------------------------------------------------------------------------
 
-
-instance Show Subst where
-  show = showFix
-
-instance Fixpoint Subst where
-  toFix (Su m) = case hashMapToAscList m of
-                   []  -> empty
-                   xys -> hcat $ map (\(x,y) -> brackets $ toFix x <> text ":=" <> toFix y) xys
-
-targetSubstSyms :: Subst -> [Symbol]
-targetSubstSyms (Su ms) = syms $ M.elems ms
-
-instance PPrint Reft where
-  pprint r@(Reft (_,ra))
-    | isTauto r        = text "true"
-    | otherwise        = {- intersperse comma -} pprintBin z trueD andD flat
-    where
-      flat = flattenRefas [ra]
-      z    = if length flat > 1 then 3 else 0
-
-instance PPrint SortedReft where
-  pprint (RR so (Reft (v, ras)))
-    = braces
-    $ pprint v <+> text ":" <+> toFix so <+> text "|" <+> pprint ras
-
-instance PPrint a => PPrint (Located a) where
-  pprint (Loc _ _ x) = pprint x
-
-
 instance PTable (SInfo a) where
   ptable fi = DocTable [ (text "# Sub Constraints", pprint $ length $ cm fi)
                        , (text "# WF  Constraints", pprint $ length $ ws fi)
                        ]
-
---------------------------------------------------------------------------------
--- | Predicates ----------------------------------------------------------------
---------------------------------------------------------------------------------
-
-
-instance Hashable Brel
-instance Hashable Bop
-instance Hashable SymConst
-instance Hashable Constant
-
-instance Fixpoint Pred where
-  toFix PTop             = text "???"
-  toFix PTrue            = text "true"
-  toFix PFalse           = text "false"
-  toFix (PBexp e)        = parens $ text "?" <+> toFix e
-  toFix (PNot p)         = parens $ text "~" <+> parens (toFix p)
-  toFix (PImp p1 p2)     = parens $ toFix p1 <+> text "=>" <+> toFix p2
-  toFix (PIff p1 p2)     = parens $ toFix p1 <+> text "<=>" <+> toFix p2
-  toFix (PAnd ps)        = text "&&" <+> toFix ps
-  toFix (POr  ps)        = text "||" <+> toFix ps
-  toFix (PAtom r e1 e2)  = parens $ toFix e1 <+> toFix r <+> toFix e2
-  toFix (PKVar k su)     = toFix k <> toFix su
-  toFix (PAll xts p)     = text "forall" <+> toFix xts <+> text "." <+> toFix p
-  toFix (PExist xts p)   = text "exists" <+> toFix xts <+> text "." <+> toFix p
-
-  simplify (PAnd [])     = PTrue
-  simplify (POr  [])     = PFalse
-  simplify (PAnd [p])    = simplify p
-  simplify (POr  [p])    = simplify p
-
-  simplify (PAnd ps)
-    | any isContraPred ps = PFalse
-    | otherwise           = PAnd $ filter (not . isTautoPred) $ map simplify ps
-
-  simplify (POr  ps)
-    | any isTautoPred ps = PTrue
-    | otherwise          = POr  $ filter (not . isContraPred) $ map simplify ps
-
-  simplify p
-    | isContraPred p     = PFalse
-    | isTautoPred  p     = PTrue
-    | otherwise          = p
-
-isContraPred   :: Pred -> Bool
-isContraPred z = eqC z || (z `elem` contras)
-  where
-    contras    = [PFalse]
-
-    eqC (PAtom Eq (ECon x) (ECon y))
-               = x /= y
-    eqC (PAtom Ueq (ECon x) (ECon y))
-               = x /= y
-    eqC (PAtom Ne x y)
-               = x == y
-    eqC (PAtom Une x y)
-               = x == y
-    eqC _      = False
-
-isTautoPred   :: Pred -> Bool
-isTautoPred z  = z == PTop || z == PTrue || eqT z
-  where
-    eqT (PAnd [])
-               = True
-    eqT (PAtom Le x y)
-               = x == y
-    eqT (PAtom Ge x y)
-               = x == y
-    eqT (PAtom Eq x y)
-               = x == y
-    eqT (PAtom Ueq x y)
-               = x == y
-    eqT (PAtom Ne (ECon x) (ECon y))
-               = x /= y
-    eqT (PAtom Une (ECon x) (ECon y))
-               = x /= y
-    eqT _      = False
-
-isEVar :: Expr -> Bool
-isEVar (EVar _) = True
-isEVar _        = False
-
-isEq  :: Brel -> Bool
-isEq r          = r == Eq || r == Ueq
-
-
-isSingletonReft :: Reft -> Maybe Expr
-isSingletonReft (Reft (v, ra)) = firstMaybe (isSingletonExpr v) $ conjuncts ra
-
-firstMaybe :: (a -> Maybe b) -> [a] -> Maybe b
-firstMaybe f = listToMaybe . mapMaybe f
-
---   where
---     go (PAtom r e1 e2) | e1 == EVar v && isEq r = Just e2
---                        | e2 == EVar v && isEq r = Just e1
---     go _                                        = Nothing
-
-isSingletonExpr :: Symbol -> Pred -> Maybe Expr
-isSingletonExpr v (PAtom r e1 e2)
-  | e1 == EVar v && isEq r = Just e2
-  | e2 == EVar v && isEq r = Just e1
-isSingletonExpr _ _        = Nothing
-
-pAnd, pOr     :: ListNE Pred -> Pred
-pAnd          = simplify . PAnd
-pOr           = simplify . POr
-pIte p1 p2 p3 = pAnd [p1 `PImp` p2, (PNot p1) `PImp` p3]
-
-mkProp        = PBexp . EApp (dummyLoc propConName) . (: [])
-
-pprReft (Reft (v, p)) d
-  | isTautoPred p
-  = d
-  | otherwise
-  = braces (toFix v <+> colon <+> d <+> text "|" <+> ppRas [p])
-
-pprReftPred (Reft (_, p))
-  | isTautoPred p
-  = text "true"
-  | otherwise
-  = ppRas [p]
-
-ppRas = cat . punctuate comma . map toFix . flattenRefas
-
-------------------------------------------------------------------------
--- | Generalizing Symbol, Expression, Predicate into Classes -----------
-------------------------------------------------------------------------
-
--- | Values that can be viewed as Constants
-
--- | Values that can be viewed as Expressions
-
-class Expression a where
-  expr   :: a -> Expr
-
--- | Values that can be viewed as Predicates
-
-class Predicate a where
-  prop   :: a -> Pred
-
-instance Expression Expr where
-  expr = id
-
--- | The symbol may be an encoding of a SymConst.
-
-instance Expression Symbol where
-  expr s = maybe (eVar s) ESym (decodeSymConst s)
-  -- expr = eVar
-
-instance Expression Text where
-  expr = ESym . SL
-
-instance Expression Integer where
-  expr = ECon . I
-
-instance Expression Int where
-  expr = expr . toInteger
-
-instance Predicate Symbol where
-  prop = eProp
-
-instance Predicate Pred where
-  prop = id
-
-instance Predicate Bool where
-  prop True  = PTrue
-  prop False = PFalse
-
-eVar ::  Symbolic a => a -> Expr
-eVar = EVar . symbol
-
-eProp ::  Symbolic a => a -> Pred
-eProp = mkProp . eVar
-
-relReft :: (Expression a) => Brel -> a -> Reft
-relReft r e   = Reft (vv_, PAtom r (eVar vv_)  (expr e))
-
-exprReft, notExprReft, uexprReft ::  (Expression a) => a -> Reft
-exprReft      = relReft Eq
-notExprReft   = relReft Ne
-uexprReft     = relReft Ueq
-
-propReft      ::  (Predicate a) => a -> Reft
-propReft p    = Reft (vv_, PIff (eProp vv_) (prop p))
-
-predReft      :: (Predicate a) => a -> Reft
-predReft p    = Reft (vv_, prop p)
-
-reft :: Symbol -> Pred -> Reft
-reft v p = Reft (v, p)
-
-mapPredReft :: (Pred -> Pred) -> Reft -> Reft
-mapPredReft f (Reft (v, p)) = Reft (v, f p)
-
-
----------------------------------------------------------------
------------------ Refinements ---------------------------------
----------------------------------------------------------------
-
-newtype Reft = Reft (Symbol, Pred)
-               deriving (Eq, Data, Typeable, Generic)
-
-instance Show Reft where
-  show (Reft x) = render $ toFix x
-
-data SortedReft = RR { sr_sort :: !Sort, sr_reft :: !Reft }
-                  deriving (Eq, Show, Data, Typeable, Generic)
-
-isFunctionSortedReft :: SortedReft -> Bool
-isFunctionSortedReft = isJust . functionSort . sr_sort
-
-functionSort :: Sort -> Maybe (Int, [Sort], Sort)
-functionSort (FFunc n ts) = Just (n, its, t)
-  where
-    (its, t)              = safeUnsnoc "functionSort" ts
-functionSort _            = Nothing
-
-
-isNonTrivial :: Reftable r => r -> Bool
-isNonTrivial = not . isTauto
-
-reftPred :: Reft -> Pred
-reftPred (Reft (_, p)) = p
-
-reftBind :: Reft -> Symbol
-reftBind (Reft (x, _)) = x
-
 
 ---------------------------------------------------------------
 -- | Environments ---------------------------------------------
@@ -520,21 +214,10 @@ adjustBindEnv f i (BE n m) = BE n $ M.adjust f i m
 instance Functor SEnv where
   fmap = mapSEnv
 
-instance Fixpoint Reft where
-  toFix = pprReftPred
-
-instance Fixpoint SortedReft where
-  toFix (RR so (Reft (v, ra)))
-    = braces
-    $ toFix v <+> text ":" <+> toFix so <+> text "|" <+> (toFix $ conjuncts ra)
-
 instance Fixpoint BindEnv where
   toFix (BE _ m) = vcat $ map toFixBind $ hashMapToAscList m
 
 toFixBind (i, (x, r)) = text "bind" <+> toFix i <+> toFix x <+> text ":" <+> toFix r
-
-hashMapToAscList :: Ord a => M.HashMap a b -> [(a, b)]
-hashMapToAscList = sortBy (\x y -> compare (fst x) (fst y)) . M.toList
 
 -- instance (Fixpoint a) => Fixpoint (SEnv a) where
 --   toFix (SE e)    = vcat $ map pprxt $ hashMapToAscList e
@@ -548,7 +231,7 @@ instance Fixpoint (SEnv a) => Show (SEnv a) where
   show = render . toFix
 
 -----------------------------------------------------------------------------
-------------------- Constraints ---------------------------------------------
+-- | Constraints --------------------------------------------------------------
 -----------------------------------------------------------------------------
 
 {-@ type Tag = { v : [Int] | len(v) = 1 } @-}
@@ -726,7 +409,7 @@ instance Fixpoint a => Fixpoint (WfC a) where
   toFix w     = hang (text "\n\nwf:") 2 bd
     where bd  =   toFix (wenv w)
               -- NOTE: this next line is printed this way for compatability with the OCAML solver
-              $+$ text "reft" <+> toFix (RR t (Reft (v, PKVar k emptySubst)))
+              $+$ text "reft" <+> toFix (RR t (Reft (v, PKVar k mempty)))
               $+$ toFixMeta (text "wf") (toFix (winfo w))
           (v, t, k) = wrft w
 
@@ -737,75 +420,22 @@ toFixMeta k v = text "// META" <+> k <+> text ":" <+> v
 pprId (Just i)  = text "id" <+> tshow i
 pprId _         = text ""
 
-------------------------------------------------------------
-------------- Generally Useful Refinements -----------------
-------------------------------------------------------------
-
-symbolReft    :: (Symbolic a) => a -> Reft
-symbolReft    = exprReft . eVar
-
-usymbolReft   :: (Symbolic a) => a -> Reft
-usymbolReft   = uexprReft . eVar
-
-vv_ :: Symbol
-vv_ = vv Nothing
-
-trueSortedReft :: Sort -> SortedReft
-trueSortedReft = (`RR` trueReft)
-
-trueReft, falseReft :: Reft
-trueReft  = Reft (vv_, PTrue)
-falseReft = Reft (vv_, PFalse)
-
-flattenRefas :: [Pred] -> [Pred]
-flattenRefas        = concatMap flatP
-  where
-    flatP (PAnd ps) = concatMap flatP ps
-    flatP p         = [p]
-
---squishRefas     :: [Refa] -> [Refa]
---squishRefas ras =  [squish (raPred <$> ras)]
---  where
---    squish      = Refa . pAnd . sortNub . filter (not . isTautoPred) . concatMap conjuncts
-
-conjuncts :: Pred -> [Pred]
-conjuncts (PAnd ps) = concatMap conjuncts ps
-conjuncts p
-  | isTautoPred p   = []
-  | otherwise       = [p]
 
 ----------------------------------------------------------------
 -- | Serialization ---------------------------------------------
 ----------------------------------------------------------------
 
-
-
 instance (Hashable a, Eq a, B.Binary a) => B.Binary (S.HashSet a) where
   put = B.put . S.toList
   get = S.fromList <$> B.get
 
-instance (Hashable k, Eq k, B.Binary k, B.Binary v) => B.Binary (M.HashMap k v) where
-  put = B.put . M.toList
-  get = M.fromList <$> B.get
-
-instance B.Binary KVar
-instance B.Binary Kuts
 instance B.Binary Qualifier
-
-
-instance B.Binary Subst
+instance B.Binary Kuts
 instance B.Binary IBindEnv
 instance B.Binary BindEnv
-instance B.Binary Constant
-instance B.Binary SymConst
-instance B.Binary Brel
-instance B.Binary Bop
-instance B.Binary Expr
-instance B.Binary Pred
-instance B.Binary Reft
-instance B.Binary SortedReft
+
+
 instance (B.Binary a) => B.Binary (SEnv a)
-instance (B.Binary a) => B.Binary (FixResult a)
 instance (B.Binary a) => B.Binary (SubC a)
 instance (B.Binary a) => B.Binary (WfC a)
 instance (B.Binary a) => B.Binary (SimpC a)
@@ -815,23 +445,13 @@ instance (B.Binary (c a), B.Binary a) => B.Binary (GInfo c a)
 -- | Strictness ------------------------------------------------
 ----------------------------------------------------------------
 
-instance NFData KVar
-instance NFData Kuts
 instance NFData Qualifier
-instance NFData FTycon
-instance NFData Sort
-instance NFData Sub
-instance NFData Subst
+instance NFData Kuts
 instance NFData IBindEnv
 instance NFData BindEnv
-instance NFData Constant
-instance NFData SymConst
-instance NFData Brel
-instance NFData Bop
-instance NFData Expr
-instance NFData Pred
-instance NFData Reft
-instance NFData SortedReft
+
+
+
 instance (NFData a) => NFData (SEnv a)
 instance (NFData a) => NFData (SubC a)
 instance (NFData a) => NFData (WfC a)
@@ -1018,154 +638,8 @@ writeFInfo :: (Fixpoint a, Fixpoint (c a)) => Config -> GInfo c a -> FilePath ->
 writeFInfo cfg fi f = writeFile f (render $ toFixpoint cfg fi)
 
 -------------------------------------------------------------------------
--- | A Class Predicates for Valid Refinements Types ---------------------
+-- | Class Predicates for Valid Refinements -----------------------------
 -------------------------------------------------------------------------
-
-class (Monoid r, Subable r) => Reftable r where
-  isTauto :: r -> Bool
-  ppTy    :: r -> Doc -> Doc
-
-  top     :: r -> r
-  top _   =  mempty
-
-  bot     :: r -> r
-
-  meet    :: r -> r -> r
-  meet    = mappend
-
-  toReft  :: r -> Reft
-  ofReft  :: Reft -> r
-  params  :: r -> [Symbol]          -- ^ parameters for Reft, vv + others
-
-instance Monoid Pred where
-  mempty      = PTrue
-  mappend p q = pAnd [p, q]
-  mconcat     = pAnd
-
-instance Monoid Reft where
-  mempty  = trueReft
-  mappend = meetReft
-
-meetReft (Reft (v, ra)) (Reft (v', ra'))
-  | v == v'          = Reft (v , ra  `mappend` ra')
-  | v == dummySymbol = Reft (v', ra' `mappend` (ra `subst1`  (v , EVar v')))
-  | otherwise        = Reft (v , ra  `mappend` (ra' `subst1` (v', EVar v )))
-
-instance Subable Reft where
-  syms (Reft (v, ras))      = v : syms ras
-  substa f (Reft (v, ras))  = Reft (f v, substa f ras)
-  subst su (Reft (v, ras))  = Reft (v, subst (substExcept su [v]) ras)
-  substf f (Reft (v, ras))  = Reft (v, substf (substfExcept f [v]) ras)
-  subst1 (Reft (v, ras)) su = Reft (v, subst1Except [v] ras su)
-
-
-instance Subable SortedReft where
-  syms               = syms . sr_reft
-  subst su (RR so r) = RR so $ subst su r
-  substf f (RR so r) = RR so $ substf f r
-  substa f (RR so r) = RR so $ substa f r
-
-instance Reftable () where
-  isTauto _ = True
-  ppTy _  d = d
-  top  _    = ()
-  bot  _    = ()
-  meet _ _  = ()
-  toReft _  = mempty
-  ofReft _  = mempty
-  params _  = []
-
-instance Reftable Reft where
-  isTauto  = all isTautoPred . conjuncts . reftPred
-  ppTy     = pprReft
-  toReft   = id
-  ofReft   = id
-  params _ = []
-
-  bot    _        = falseReft
-  top (Reft(v,_)) = Reft (v, mempty)
-
-instance Monoid Sort where
-  mempty            = FObj "any"
-  mappend t1 t2
-    | t1 == mempty  = t2
-    | t2 == mempty  = t1
-    | t1 == t2      = t1
-    | otherwise     = errorstar $ "mappend-sort: conflicting sorts t1 =" ++ show t1 ++ " t2 = " ++ show t2
-
-instance Monoid SortedReft where
-  mempty        = RR mempty mempty
-  mappend t1 t2 = RR (mappend (sr_sort t1) (sr_sort t2)) (mappend (sr_reft t1) (sr_reft t2))
-
-instance Reftable SortedReft where
-  isTauto  = isTauto . toReft
-  ppTy     = ppTy . toReft
-  toReft   = sr_reft
-  ofReft   = errorstar "No instance of ofReft for SortedReft"
-  params _ = []
-  bot s    = s { sr_reft = falseReft }
-
-class Falseable a where
-  isFalse :: a -> Bool
-
-instance Falseable Pred where
-  isFalse (PFalse) = True
-  isFalse _        = False
-
-instance Falseable Reft where
-  isFalse (Reft (_, ra)) = isFalse ra
-
----------------------------------------------------------------
--- | String Constants -----------------------------------------
----------------------------------------------------------------
-
--- | Replace all symbol-representations-of-string-literals with string-literal
---   Used to transform parsed output from fixpoint back into fq.
-
-
-instance Symbolic SymConst where
-  symbol = encodeSymConst
-
-encodeSymConst        :: SymConst -> Symbol
-encodeSymConst (SL s) = litPrefix `mappend` symbol s
-
-decodeSymConst :: Symbol -> Maybe SymConst
-decodeSymConst = fmap (SL . symbolText) . stripPrefix litPrefix
-
----------------------------------------------------------------
--- | Edit Distance --------------------------------------------
----------------------------------------------------------------
-
-
-editDistance :: Eq a => [a] -> [a] -> Int
-editDistance xs ys = table ! (m,n)
-    where
-    (m,n) = (length xs, length ys)
-    x     = array (1,m) (zip [1..] xs)
-    y     = array (1,n) (zip [1..] ys)
-
-    table :: Array (Int,Int) Int
-    table = array bnds [(ij, dist ij) | ij <- range bnds]
-    bnds  = ((0,0),(m,n))
-
-    dist (0,j) = j
-    dist (i,0) = i
-    dist (i,j) = minimum [table ! (i-1,j) + 1, table ! (i,j-1) + 1,
-        if x ! i == y ! j then table ! (i-1,j-1) else 1 + table ! (i-1,j-1)]
-
------------------------------------------------------------------------------
--- | Located Values ---------------------------------------------------------
------------------------------------------------------------------------------
-
-instance Expression a => Expression (Located a) where
-  expr   = expr . val
-
-instance Subable a => Subable (Located a) where
-  syms (Loc _ _ x)   = syms x
-  substa f (Loc l l' x) = Loc l l' (substa f x)
-  substf f (Loc l l' x) = Loc l l' (substf f x)
-  subst su (Loc l l' x) = Loc l l' (subst su x)
-
 ---------------------------------------------------------------------------
 -- | FInfo to SInfo conversion
 ---------------------------------------------------------------------------
